Better tests for the generic notifier
