+++ lib/stoplight/rspec.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require 'stoplight/rspec/generic_notifier'
