+++ spec/stoplight/notifier/generic_spec.rb
@@ -2,59 +2,19 @@
 
 require 'spec_helper'
 
-RSpec.shared_examples_for 'a generic notifier' do
-  it 'includes Generic' do
-    expect(described_class).to include(Stoplight::Notifier::Generic)
-  end
-
-  describe '#formatter' do
-    it 'is initially the default' do
-      formatter = nil
-      expect(described_class.new(nil, formatter).formatter)
-        .to eql(Stoplight::Default::FORMATTER)
-    end
-
-    it 'reads the formatter' do
-      formatter = proc {}
-      expect(described_class.new(nil, formatter).formatter)
-        .to eql(formatter)
-    end
-  end
-
-  describe '#notify' do
-    let(:light) { Stoplight::Light.new(name, &code) }
-    let(:name) { ('a'..'z').to_a.shuffle.join }
-    let(:code) { -> {} }
-    let(:from_color) { Stoplight::Color::GREEN }
-    let(:to_color) { Stoplight::Color::RED }
-    let(:notifier) { described_class.new(double.as_null_object) }
-
-    it 'returns the message' do
-      error = nil
-      expect(notifier.notify(light, from_color, to_color, error))
-        .to eql(notifier.formatter.call(light, from_color, to_color, error))
-    end
-
-    it 'returns the message with an error' do
-      error = ZeroDivisionError.new('divided by 0')
-      expect(notifier.notify(light, from_color, to_color, error))
-        .to eql(notifier.formatter.call(light, from_color, to_color, error))
-    end
-  end
-end
-
 RSpec.describe Stoplight::Notifier::Generic do
+  let(:light) { Stoplight::Light.new(name) {} }
+  let(:name) { ('a'..'z').to_a.shuffle.join }
+  let(:from_color) { Stoplight::Color::GREEN }
+  let(:to_color) { Stoplight::Color::RED }
+  let(:error) { nil }
+
   it 'is a module' do
     expect(described_class).to be_a(Module)
   end
 
-  describe '#notify' do
-    let(:light) { Stoplight::Light.new(name) {} }
-    let(:name) { ('a'..'z').to_a.shuffle.join }
-    let(:from_color) { Stoplight::Color::GREEN }
-    let(:to_color) { Stoplight::Color::RED }
+  describe '#put' do
     let(:notifier) { notifier_class.new(double.as_null_object) }
-    let(:error) { nil }
     let(:notifier_class) do
       Class.new do
         include Stoplight::Notifier::Generic
@@ -65,6 +25,26 @@ RSpec.describe Stoplight::Notifier::Generic do
       expect do
         notifier.notify(light, from_color, to_color, error)
       end.to raise_error(NotImplementedError)
+    end
+  end
+
+  describe '#notify' do
+    let(:formatted_message) { 'formatted message' }
+    let(:notifier) { notifier_class.new(object) }
+    let(:notifier_class) do
+      Class.new do
+        include Stoplight::Notifier::Generic
+        def put(message)
+          object.put(message)
+        end
+      end
+
+      it 'puts formatted message' do
+        expect(formatter).to receive(:call).with(light, from_color, to_color, error) { formatted_message }
+        expect(object).to receive(:put).with(formatted_message)
+
+        expect(notifier.notify(light, from_color, to_color, error)).to eq(formatted_message)
+      end
     end
   end
 end
